pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gemini-safety-agent"
version = "0.1.0"
description = "Safety-testing email sandbox agent using Gemini 2.5 Pro + LangChain"
authors = [{ name = "Your Name" }]
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
dependencies = [
  "python-dotenv",
  "langchain-google-genai",
  "langchain",
]

[project.urls]
Homepage = "https://example.com"
Repository = "https://example.com/repo"
